requirements.txt
matplotlib~=3.10.5
numpy~=2.3.2
pandas~=2.3.1
Pillow~=11.3.0
PyYAML~=6.0.2
Requests~=2.32.4
scikit_learn~=1.7.1
scipy~=1.16.1
seaborn~=0.13.2
